Add channel force-closing support
In order to support this, this commit adds two elements:

1) CommitmentToLocalExtension

This is an NBitcoin BuilderExtension that tells
NBitcoin.TransactionBuilder how to recognise and sign to_local txouts
from lightning commitment transactions. This is needed for force-closing
to spend to_local outputs.

2) tryGetFundsFromLocalCommitmentTx

When force-closing a channel this function is used to recover funds from
our commitment transaction. It generates a transaction which can be
broadcast once the timelock on the to_local output has expired.